+++ RogueBasin/RogueBasin/Creatures/MaintBot.cs
@@ -179,7 +179,7 @@ namespace RogueBasin.Creatures
         /// <summary>
         /// Creature damage modifier.  Set by type of creature.
         /// </summary>
-        public override int DamageModifier()
+        public override double DamageModifier()
         {
             return 1;
         }
