+++ RogueBasin/RogueBasin/Items/BioWare.cs
@@ -89,7 +89,7 @@ namespace RogueBasin.Items
             return 0;
         }
 
-        public int DamageModifier()
+        public double DamageModifier()
         {
             return 0;
         }
