+++ RogueBasin/RogueBasin/Items/StealthCloak.cs
@@ -104,7 +104,7 @@ namespace RogueBasin.Items
             return 0;
         }
 
-        public int DamageModifier()
+        public double DamageModifier()
         {
             return 0;
         }
