+++ RogueBasin/RogueBasin/Items/StunGrenade.cs
@@ -195,7 +195,7 @@ namespace RogueBasin.Items
             return 0;
         }
 
-        public int DamageModifier()
+        public double DamageModifier()
         {
             return 0;
         }
