+++ RogueBasin/RogueBasin/Items/Vibroblade.cs
@@ -117,7 +117,7 @@ namespace RogueBasin.Items
             return 0;
         }
 
-        public int DamageModifier()
+        public double DamageModifier()
         {
             return 0;
         }
